# Improve performance of circuit_to_instruction for circuits without parameters