--- a/qiskit/converters/circuit_to_instruction.py
+++ b/qiskit/converters/circuit_to_instruction.py
@@ -83,9 +83,12 @@ def circuit_to_instruction(circuit, parameter_map=None, equivalence_library=None
     )
     out_instruction.condition = None
 
-    target = circuit.assign_parameters(parameter_dict, inplace=False)
-
-    if equivalence_library is not None:
+    if equivalence_library is None and not parameter_dict:
+        # avoid making a copy of the circuit
+        # parameter_dict is empty and we do not have to pass a copy to the equivalence library
+        target = circuit 
+    else:
+        target = circuit.assign_parameters(parameter_dict, inplace=False)
         equivalence_library.add_equivalence(out_instruction, target)
 
     regs = []