<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [x] I have added the tests to cover my changes.
- [x] I have updated the documentation accordingly.
- [x] I have read the CONTRIBUTING document.
-->

### Summary

The performance of `circuit_to_instruction` is important for many operations involving the `QuantumCircuit`. We improve performance by avoiding an internal copy of the circuit when no parameters are present.


### Details and comments

Benchmark:
```
from qiskit.circuit import QuantumCircuit

qc=QuantumCircuit(1)
qc.h(0)
qc.x(0)
%timeit qc.to_instruction()
```

```
Main: 127 µs ± 16.1 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
PR: 68.1 µs ± 9.11 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```