@jakelishman 2023-09-05
If you're relying heavily on the performance of this method, you'll likely find it faster to build up custom `Instruction` instances yourself directly - you'll be able to skip _all_ the copying / rewriting by building up the definition in the correct form from the start, and then not mutating it.  For example, you can do
```python
from qiskit.circuit import QuantumCircuit, Instruction

defn = QuantumCircuit(1)
defn.h(0)
defn.x(0)
instr = Instruction("my_instruction", 1, 0, [])
instr.definition = defn
```
and it'll have zero cost other than the initial construction of the circuit.

You'll not be able to do that if you
- can't rely on the definition circuit being built with flat registers
- can't rely on the definition circuit not containing conditionals
- need to mutate the definition circuit after you've created it (it'd also modify the instruction).

All that said, there's no harm in accepting the PR.  We could consider going further, and give `QuantumCircuit.to_instruction` and `circuit_to_instruction` (and similar for `to_gate`) a `copy` argument that can be set to `False` to do the simple definition-bind I described above, at the cost of taking ownership of the circuit object.