@eendebakpt 2023-09-05
@jakelishman The exact location where the `circuit_to_instruction()` is used, is https://github.com/Qiskit-Extensions/qiskit-experiments/blob/6a06e7489dba5a84ae263d134622962d906beaf7/qiskit_experiments/library/randomized_benchmarking/clifford_utils.py#L140

A copy argument would work even better, as (needs to be checked) we can borrow the circuit argument. There would have to be a check on the conditionals and flat registers though (not sure how to do that)