@jakelishman 2023-09-05
That code in Experiments could probably be made a whole bunch faster on their side.  It's worth considering whether they want to do:
```python
@lru_cache(maxsize=11520)
def _clifford_2q_int_to_gate(
    num: Integral, basis_gates: Optional[Tuple[str]]
) -> Gate:
    out = Gate(f"clifford2q_{num}", 2, [])
    out.definition = CliffordUtils.clifford_2_qubit_circuit(num, basis_gates)
    return out
```
which has eager synthesis and is cached at the module level, or whether they'd prefer to re-apply the `@lru_cache` decorator to `CliffordUtils.cliff_2_qubit_circuit` and instead do
```python
def _clifford_2q_int_to_gate(num, basis_gates):
    out = Gate(f"clifford2q_{num}", 2, [])
    out._define = functools.partial(CliffordUtils.clifford_2_qubit_circuit(num, basis_gates))
    return out
```
which has lazy but cached synthesis.  Both would avoid all cost from conversion to instruction (though in the second case, you might want to check I got the pickling behaviour correct).

Again, that's not to say we shouldn't also improve `circuit_to_instruction` and `circuit_to_gate` on our side as well.